feat: Enable dark mode by default
Set the `lang` attribute to "en" and add the "dark" class to the html tag in index.html. This ensures that the application loads with dark mode enabled by default, improving the user experience for new visitors. The explicit JavaScript to add the dark class has been removed as it is no longer necessary.